Cursor indices are long. Prepend the GTK+ generated expose if we already
2007-01-19  Chris Wilson <chris@chris-wilson.co.uk>

	* src/vte-private.h:
	* src/vte.c: (_vte_invalidate_cell):
		Cursor indices are long.
	* src/vte.c: (_vte_terminal_expose):
		Prepend the GTK+ generated expose if we already have an
		display update pending.


svn path=/trunk/; revision=1472

+++ src/vte-private.h
@@ -362,7 +362,7 @@ void _vte_invalidate_all(VteTerminal *terminal);
 void _vte_invalidate_cells(VteTerminal *terminal,
 			   glong column_start, gint column_count,
 			   glong row_start, gint row_count);
-void _vte_invalidate_cell(VteTerminal *terminal, gint col, gint row);
+void _vte_invalidate_cell(VteTerminal *terminal, glong col, glong row);
 void _vte_invalidate_cursor_once(VteTerminal *terminal, gboolean periodic);
 VteRowData * _vte_new_row_data(VteTerminal *terminal);
 VteRowData * _vte_new_row_data_sized(VteTerminal *terminal, gboolean fill);
